Reduce usage of cipher inits
