+++ src/main/java/org/opensearch/index/store/cipher/MemorySegmentDecryptor.java
@@ -4,15 +4,13 @@
  */
 package org.opensearch.index.store.cipher;
 
+import static org.opensearch.index.store.cipher.AesCipherFactory.CIPHER_POOL;
+
 import java.lang.foreign.Arena;
 import java.lang.foreign.MemorySegment;
 import java.nio.ByteBuffer;
-import java.security.NoSuchAlgorithmException;
-import java.security.NoSuchProviderException;
-import java.util.Arrays;
 
 import javax.crypto.Cipher;
-import javax.crypto.NoSuchPaddingException;
 import javax.crypto.spec.IvParameterSpec;
 import javax.crypto.spec.SecretKeySpec;
 
@@ -28,14 +26,6 @@ public class MemorySegmentDecryptor {
     private static final byte[] ZERO_SKIP = new byte[AesCipherFactory.AES_BLOCK_SIZE_BYTES];
     private static final int DEFAULT_MAX_CHUNK_SIZE = 16_384;
 
-    private static final ThreadLocal<Cipher> CIPHER_POOL = ThreadLocal.withInitial(() -> {
-        try {
-            return Cipher.getInstance("AES/CTR/NoPadding", "SunJCE");
-        } catch (NoSuchAlgorithmException | NoSuchProviderException | NoSuchPaddingException e) {
-            throw new RuntimeException(e);
-        }
-    });
-
     private MemorySegmentDecryptor() {
 
     }
@@ -43,15 +33,8 @@ public class MemorySegmentDecryptor {
     public static void decryptInPlace(Arena arena, long addr, long length, byte[] key, byte[] iv, long fileOffset) throws Exception {
         // Get thread-local cipher
         Cipher cipher = CIPHER_POOL.get();
-        SecretKeySpec keySpec = new SecretKeySpec(key, "AES");
-        byte[] ivCopy = Arrays.copyOf(iv, iv.length);
-
-        int blockOffset = (int) (fileOffset / AesCipherFactory.AES_BLOCK_SIZE_BYTES);
-
-        ivCopy[AesCipherFactory.IV_ARRAY_LENGTH - 1] = (byte) blockOffset;
-        ivCopy[AesCipherFactory.IV_ARRAY_LENGTH - 2] = (byte) (blockOffset >>> 8);
-        ivCopy[AesCipherFactory.IV_ARRAY_LENGTH - 3] = (byte) (blockOffset >>> 16);
-        ivCopy[AesCipherFactory.IV_ARRAY_LENGTH - 4] = (byte) (blockOffset >>> 24);
+        SecretKeySpec keySpec = new SecretKeySpec(key, AesCipherFactory.ALGORITHM);
+        byte[] ivCopy = AesCipherFactory.computeOffsetIV(iv, fileOffset);
 
         cipher.init(Cipher.DECRYPT_MODE, keySpec, new IvParameterSpec(ivCopy));
 
@@ -91,15 +74,8 @@ public class MemorySegmentDecryptor {
 
     public static void decryptInPlace(long addr, long length, byte[] key, byte[] iv, long fileOffset) throws Exception {
         Cipher cipher = CIPHER_POOL.get();
-        SecretKeySpec keySpec = new SecretKeySpec(key, "AES");
-        byte[] ivCopy = Arrays.copyOf(iv, iv.length);
-
-        int blockOffset = (int) (fileOffset / AesCipherFactory.AES_BLOCK_SIZE_BYTES);
-
-        ivCopy[AesCipherFactory.IV_ARRAY_LENGTH - 1] = (byte) blockOffset;
-        ivCopy[AesCipherFactory.IV_ARRAY_LENGTH - 2] = (byte) (blockOffset >>> 8);
-        ivCopy[AesCipherFactory.IV_ARRAY_LENGTH - 3] = (byte) (blockOffset >>> 16);
-        ivCopy[AesCipherFactory.IV_ARRAY_LENGTH - 4] = (byte) (blockOffset >>> 24);
+        SecretKeySpec keySpec = new SecretKeySpec(key, AesCipherFactory.ALGORITHM);
+        byte[] ivCopy = AesCipherFactory.computeOffsetIV(iv, fileOffset);
 
         cipher.init(Cipher.DECRYPT_MODE, keySpec, new IvParameterSpec(ivCopy));
 
@@ -139,15 +115,8 @@ public class MemorySegmentDecryptor {
 
     public static void decryptSegment(MemorySegment segment, long fileOffset, byte[] key, byte[] iv, int segmentSize) throws Exception {
         Cipher cipher = CIPHER_POOL.get();
-        SecretKeySpec keySpec = new SecretKeySpec(key, "AES");
-        byte[] ivCopy = Arrays.copyOf(iv, iv.length);
-
-        int blockOffset = (int) (fileOffset / AesCipherFactory.AES_BLOCK_SIZE_BYTES);
-
-        ivCopy[AesCipherFactory.IV_ARRAY_LENGTH - 1] = (byte) blockOffset;
-        ivCopy[AesCipherFactory.IV_ARRAY_LENGTH - 2] = (byte) (blockOffset >>> 8);
-        ivCopy[AesCipherFactory.IV_ARRAY_LENGTH - 3] = (byte) (blockOffset >>> 16);
-        ivCopy[AesCipherFactory.IV_ARRAY_LENGTH - 4] = (byte) (blockOffset >>> 24);
+        SecretKeySpec keySpec = new SecretKeySpec(key, AesCipherFactory.ALGORITHM);
+        byte[] ivCopy = AesCipherFactory.computeOffsetIV(iv, fileOffset);
 
         cipher.init(Cipher.DECRYPT_MODE, keySpec, new IvParameterSpec(ivCopy));
