+++ src/main/java/org/opensearch/index/store/niofs/CryptoNIOFSDirectory.java
@@ -13,15 +13,12 @@ import java.nio.file.StandardOpenOption;
 import java.security.Provider;
 import java.util.concurrent.atomic.AtomicLong;
 
-import javax.crypto.Cipher;
-
 import org.apache.lucene.store.IOContext;
 import org.apache.lucene.store.IndexInput;
 import org.apache.lucene.store.IndexOutput;
 import org.apache.lucene.store.LockFactory;
 import org.apache.lucene.store.NIOFSDirectory;
 import org.opensearch.common.util.io.IOUtils;
-import org.opensearch.index.store.cipher.AesCipherFactory;
 import org.opensearch.index.store.iv.KeyIvResolver;
 
 /**
@@ -54,14 +51,10 @@ public class CryptoNIOFSDirectory extends NIOFSDirectory {
         boolean success = false;
 
         try {
-            Cipher cipher = AesCipherFactory.getCipher(provider);
-            AesCipherFactory.initCipher(cipher, keyIvResolver.getDataKey(), keyIvResolver.getIvBytes(), Cipher.DECRYPT_MODE, 0);
-
             final IndexInput indexInput = new CryptoBufferedIndexInput(
                 "CryptoBufferedIndexInput(path=\"" + path + "\")",
                 fc,
                 context,
-                cipher,
                 this.keyIvResolver
             );
             success = true;
