+++ src/test/java/org/opensearch/index/store/mmap/LazyDecryptedRaceConditionTests.java
@@ -4,24 +4,29 @@
  */
 package org.opensearch.index.store.mmap;
 
-import static org.junit.Assert.*;
-
 import java.lang.foreign.Arena;
 import java.lang.foreign.MemorySegment;
 import java.lang.foreign.ValueLayout;
 import java.util.Arrays;
-import java.util.concurrent.*;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 
 import org.junit.After;
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 import org.junit.Before;
 import org.junit.Test;
+import org.opensearch.common.SuppressForbidden;
 
 /**
  * The test focuses on the specific scenario where multiple
  * threads attempt to read the same encrypted page simultaneously.
  */
+@SuppressWarnings("preview")
+@SuppressForbidden(reason = "temporary bypass")
 public class LazyDecryptedRaceConditionTests {
 
     // Use system page size aligned values to avoid boundary issues
